+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Added DMC-4K-HD and DMC-4KZ-HD input cards
+
 ## [14.1.0] - 2019-01-29
 ### Added
  - DMC-SO-HD DM 8G Fiber Output Card